Include built files, so readthedocs can finish building

+++ cgsv/projectaudio.py
@@ -2,12 +2,16 @@ import os
 from textwrap import dedent
 
 from docutils.parsers.rst.directives import unchanged_required
-import numpy as np
-from scipy.io import wavfile
 from sunvox.api import Slot
 from sunvox.buffered import BufferedProcess
 
 from .directive import Directive
+
+try:
+    import numpy as np
+    from scipy.io import wavfile
+except ImportError:
+    np = None
 
 
 class ProjectAudioDirective(Directive):
@@ -40,25 +44,26 @@ class ProjectAudioDirective(Directive):
         os.makedirs(basedir, exist_ok=True)
         filepath = os.path.join(basedir, filename)
 
-        freq = BufferedProcess.freq
-        start = int(self.options['start'] * freq)
-        duration = int(self.options['duration'] * freq)
-        fadeout = int(self.options['fadeout'] * freq)
-        length = start + duration
-        output = np.zeros((length, 2), BufferedProcess.data_type)
-        process = BufferedProcess()
-        slot = Slot(project, process=process)
-        position = 0
-        slot.play_from_beginning()
-        while position < length:
-            buffer = process.fill_buffer()
-            end_pos = min(position + BufferedProcess.size, length)
-            copy_size = end_pos - position
-            output[position:end_pos] = buffer[:copy_size]
-            position = end_pos
-        process.kill()
-        # TODO: fadeout
-        wavfile.write(filepath, freq, output)
+        if np is not None:
+            freq = BufferedProcess.freq
+            start = int(self.options['start'] * freq)
+            duration = int(self.options['duration'] * freq)
+            fadeout = int(self.options['fadeout'] * freq)
+            length = start + duration
+            output = np.zeros((length, 2), BufferedProcess.data_type)
+            process = BufferedProcess()
+            slot = Slot(project, process=process)
+            position = 0
+            slot.play_from_beginning()
+            while position < length:
+                buffer = process.fill_buffer()
+                end_pos = min(position + BufferedProcess.size, length)
+                copy_size = end_pos - position
+                output[position:end_pos] = buffer[:copy_size]
+                position = end_pos
+            process.kill()
+            # TODO: fadeout
+            wavfile.write(filepath, freq, output)
 
         download_link = dedent("""
         :download:`Download Audio for {} </_sunvox_files/{}>`
